Koduppgifter Del 1/streamlit/requirements.txt
streamlit
numpy
Pillow
tensorflow
opencv-python
scikit-learn